+++ src/metatrain/experimental/dpa3/documentation.py
@@ -77,7 +77,7 @@ with the following definitions needed to fully understand some of the parameters
 
 from typing import Literal, Optional
 
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 
 from metatrain.composition.documentation import FixedCompositionWeights
 from metatrain.utils.hypers import init_with_defaults
@@ -221,8 +221,10 @@ class ModelHypers(TypedDict):
 class TrainerHypers(TypedDict):
     """Hyperparameters for training DPA3 models."""
 
-    distributed: bool = False
-    """Whether to use distributed training."""
+    distributed: NotRequired[bool]
+    """Whether to use distributed training. When not set, distributed training
+    is enabled automatically when running under more than one SLURM task.
+    Setting this option explicitly is deprecated."""
     distributed_port: int = 39591
     """Port for DDP communication."""
     batch_size: int = 8