+++ src/metatrain/experimental/phace/documentation.py
@@ -60,7 +60,7 @@ of importance) for the PhACE architecture:
 
 from typing import Literal, Optional
 
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 
 from metatrain.composition.documentation import FixedCompositionWeights
 from metatrain.utils.hypers import init_with_defaults
@@ -192,8 +192,10 @@ class TrainerHypers(TypedDict):
     cutoffs.
     """
 
-    distributed: bool = False
-    """Whether to use distributed training."""
+    distributed: NotRequired[bool]
+    """Whether to use distributed training. When not set, distributed training
+    is enabled automatically when running under more than one SLURM task.
+    Setting this option explicitly is deprecated."""
 
     distributed_port: int = 39591
     """Port for DDP communication."""