+++ src/metatrain/llpr/trainer.py
@@ -24,7 +24,10 @@ from metatrain.utils.data import (
 from metatrain.utils.distributed.distributed_data_parallel import (
     DistributedDataParallel,
 )
-from metatrain.utils.distributed.slurm import initialize_slurm_nccl_process_group
+from metatrain.utils.distributed.slurm import (
+    initialize_slurm_nccl_process_group,
+    resolve_distributed,
+)
 from metatrain.utils.evaluate_model import evaluate_model
 from metatrain.utils.io import check_file_extension, model_from_checkpoint
 from metatrain.utils.logging import ROOT_LOGGER, MetricLogger
@@ -116,7 +119,7 @@ class Trainer(TrainerInterface[TrainerHypers]):
         if start_epoch == 0:
             model.set_wrapped_model(wrapped_model)
 
-        is_distributed = self.hypers["distributed"]
+        is_distributed = resolve_distributed(self.hypers.get("distributed"))
 
         # For the initial LLPR calibration, distributed training can be used
         if is_distributed: